Change INT to BIGINT since there might be big chanid number

+++ migrations/00003_groupid.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- SQL in this section is executed when the migration is applied.
-ALTER TABLE `standup` ADD `groupid` INT NOT NULL;
-ALTER TABLE `interns` ADD `groupid` INT NOT NULL;
+ALTER TABLE `standup` ADD `groupid` BIGINT NOT NULL;
+ALTER TABLE `interns` ADD `groupid` BIGINT NOT NULL;
 -- +goose Down
 -- SQL in this section is executed when the migration is rolled back.
 ALTER TABLE `standup` DROP `groupid`;
